EPOS Sources reorganized
