--- epos2/constants.c
@@ -1,1 +1,0 @@
-#include "constants.h"
